lbba_boosted_wsod/lbba_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
astroid=2.4.2=py36_0
blas=1.0=mkl
bzip2=1.0.8=h7b6447c_0
ca-certificates=2020.7.22=0
cairo=1.14.12=h8948797_3
certifi=2020.6.20=py36_0
cffi=1.14.0=py36h2e261b9_0
cudatoolkit=10.0.130=0
cycler=0.10.0=py36_0
dbus=1.13.14=hb2f20db_0
expat=2.2.6=he6710b0_0
ffmpeg=4.0=hcdf2ecd_0
fontconfig=2.13.0=h9420a91_0
freeglut=3.0.0=hf484d3e_5
freetype=2.9.1=h8a8886c_1
glib=2.63.1=h5a9c865_0
graphite2=1.3.13=h23475e2_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
harfbuzz=1.8.8=hffaf4a1_0
hdf5=1.10.2=hba1933b_1
icu=58.2=he6710b0_3
imageio=2.8.0=py_0
intel-openmp=2020.1=217
isort=5.5.3=py36_0
jasper=2.0.14=h07fcdf6_1
joblib=0.15.1=py_0
jpeg=9b=h024ee3a_2
kiwisolver=1.2.0=py36hfd86e86_0
lazy-object-proxy=1.4.3=py36h7b6447c_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libglu=9.0.0=hf484d3e_1
libopencv=3.4.2=hb342d67_1
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_0
libuuid=1.0.3=h1bed415_2
libvpx=1.7.0=h439df22_0
libxcb=1.13=h1bed415_1
libxml2=2.9.9=hea5a465_1
matplotlib=3.1.3=py36_0
matplotlib-base=3.1.3=py36hef1b27d_0
mccabe=0.6.1=py36_1
mkl=2020.1=217
mkl-service=2.3.0=py36he904b0f_0
mkl_fft=1.0.15=py36ha843d7b_0
mkl_random=1.1.0=py36hd6b4f25_0
ncurses=6.2=he6710b0_1
ninja=1.9.0=py36hfd86e86_0
numpy=1.18.1=py36h4f9e942_0
numpy-base=1.18.1=py36hde5b4d6_1
olefile=0.46=py_0
opencv=3.4.2=py36h6fd60c2_1
openssl=1.1.1h=h7b6447c_0
pcre=8.43=he6710b0_0
pillow=7.1.2=py36hb39fc2d_0
pip=20.0.2=py36_3
pixman=0.38.0=h7b6447c_0
py-opencv=3.4.2=py36hb342d67_1
pycparser=2.20=py_0
pylint=2.6.0=py36_0
pyparsing=2.4.7=py_0
pyqt=5.9.2=py36h05f1152_2
python=3.6.10=hcf32534_1
python-dateutil=2.8.1=py_0
pytorch=1.2.0=py3.6_cuda10.0.130_cudnn7.6.2_0
qt=5.9.7=h5867ecd_1
readline=8.0=h7b6447c_0
scikit-learn=0.22.1=py36hd81dba3_0
scipy=1.4.1=py36h0b6359f_0
setuptools=46.2.0=py36_0
sip=4.19.8=py36hf484d3e_0
six=1.14.0=py36_0
sqlite=3.31.1=h62c20be_1
tk=8.6.8=hbc83047_0
toml=0.10.1=py_0
torchvision=0.4.0=py36_cu100
tornado=6.0.4=py36h7b6447c_1
typed-ast=1.4.1=py36h7b6447c_0
wheel=0.34.2=py36_0
wrapt=1.11.2=py36h7b6447c_0
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
